requirements-postgres.txt
# PostgreSQL-specific requirements
# Install with: pip install -r requirements-postgres.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# PostgreSQL driver
psycopg2-binary==2.9.9

# Optional: asyncpg for async PostgreSQL support
# asyncpg==0.29.0